fix num ob spilled recordes

+++ mapreduce/JobInfo.h
@@ -24,25 +24,25 @@ public:
 	double compressionSize = 1.0;
 
 	//int64_t chunkSize = 128 * 1024 * 1024;
-	int64_t recordSize = 100;//the size of map record
+	double recordSize = 100;//the size of map record
 
 	int64_t mapCost = 10;//cost per record
 //	int64_t mapSize = 5;//num record * record size the same as chunk size
 	int64_t mapRecords = 50;//num record per chunk
-	int64_t mapOutRecords = 500;//num map out record per input record
-int64_t mergeCost=10;//cost per record
-	int64_t mapOutAvRecordSize = 12;//the size of map output
+	double mapOutRecords = 500;//num map out record per input record
+    int64_t mergeCost=10;//cost per record
+	double mapOutAvRecordSize = 12;//the size of map output
 	int64_t combineCost = 1;
 	int64_t combineSize = 1;//to do delete this
 //	int64_t combineRecords = 1;//to do delete
 	int64_t combineGroups = 1;//num of keys per spill
-	int64_t combineOutAvRecordSize = 12;
+	double combineOutAvRecordSize = 12;
 //	int64_t combineOutAvRecordSize_add = 0;//to do delete not used
 
 	int64_t reduceCost = 80;
 	double reduceRecords = 0;
-	int64_t reduceOutAvRecordSize = 9;
-//data
+	double reduceOutAvRecordSize = 9;
+//dataint64_t
 	DirFiles* dir = nullptr;
 	double recordsNumPerChunk = 100;
 	int numOfFiles = 10;
